+++ e2e/fixtures/citizen/caseView/contactPreferences/confirmationContent.ts
