+++ e2e/fixtures/citizen/caseView/contactPreferences/contactPreferencesContent.ts
