+++ e2e/fixtures/citizen/caseView/keepDetailsPrivate/applicant/detailsKnownContent.ts
@@ -1,0 +1,5 @@
+export enum ApplicantDetailsKnownContent {
+  applicantPageTitle = "Does the other person named in your application (the respondent) know any of your contact details?",
+  respondentPageTitle = "Do the other people named in this application (the applicants) know any of your contact details?",
+  span = "Keeping your contact details private",
+}
