+++ e2e/fixtures/citizen/caseView/keepDetailsPrivate/applicant/privateDetailsConfirmedContent.ts
@@ -1,9 +1,12 @@
 export enum ApplicantPrivateDetailsConfirmedContent {
   yesPageTitle = "The court will keep your contact details private",
   yesSpan = "Keeping your contact details private",
-  li1 = "Address",
-  li2 = "Telephone number",
-  li3 = "Email",
+  applicantLi1 = "Address",
+  applicantLi2 = "Telephone number",
+  applicantLi3 = "Email",
+  respondentLi1 = "Address",
+  respondentLi2 = "Phone Number",
+  respondentLi3 = "Email",
   h2 = "What the court will do",
   p1 = "You have told us you want to keep these contact details private:",
   p2 = "The court will hold this information securely and will not share it with anyone except Cafcass (Children and Family Court Advisory and Support Service) or Cafcass Cymru unless it is by order of the court.",
