+++ e2e/fixtures/citizen/caseView/keepDetailsPrivate/applicant/startAlternativeContent.ts
@@ -1,0 +1,11 @@
+export enum ApplicantStartAlternativeContent {
+  applicantPageTitle = "Do you want to keep your contact details private from the other person named in the application (the respondent)?",
+  respondentPageTitle = "Do you want to keep your contact details private from the other people named in the application (the applicants)?",
+  span = "Keeping your contact details private",
+  hiddenFormLabel1 = "Address",
+  hiddenFormLabel2 = "Phone number",
+  hiddenFormLabel3 = "Email",
+  hiddenDivApplicant = "Make sure you only select details the respondent does not already know.",
+  hiddenDivRespondent = "You've said that the applicants know some of your contact details. Make sure you select contact details the applicants do not already know.",
+  hiddenLegend = "Which contact details do you want to keep private from the other people in this application?",
+}
