--- e2e/fixtures/citizen/caseView/keepDetailsPrivate/detailsKnownContent.ts
@@ -1,4 +1,0 @@
-export enum ApplicantDetailsKnownContent {
-  pageTitle = "Does the other person named in your application (the respondent) know any of your contact details?",
-  span = "Keeping your contact details private",
-}
