--- e2e/fixtures/citizen/caseView/keepDetailsPrivate/startAlternativeContent.ts
@@ -1,9 +1,0 @@
-export enum ApplicantStartAlternativeContent {
-  pageTitle = "Do you want to keep your contact details private from the other person named in the application (the respondent)?",
-  span = "Keeping your contact details private",
-  hiddenFormLabel1 = "Address",
-  hiddenFormLabel2 = "Phone number",
-  hiddenFormLabel3 = "Email",
-  hiddenDiv = "Make sure you only select details the respondent does not already know.",
-  hiddenLegend = "Which contact details do you want to keep private from the other people in this application?",
-}
